requirements.txt
--extra-index-url https://pypi.ngc.nvidia.com
nvidia-tensorrt~=8.4.1.5
--extra-index-url https://download.pytorch.org/whl/cu116
torch~=1.10.2
datasets~=2.3.2
tqdm~=4.64.0
transformers~=4.19.0
